+++ lib/rubocop-grape.rb
@@ -4,9 +4,7 @@ require 'rubocop'
 
 require_relative 'rubocop/grape'
 require_relative 'rubocop/grape/version'
-require_relative 'rubocop/grape/inject'
-
-RuboCop::Grape::Inject.defaults!
+require_relative 'rubocop/grape/plugin'
 
 require_relative 'rubocop/grape/grape_files'